+++ tkviews/styles.py
@@ -1,6 +1,6 @@
 """Contains rendering steps for style nodes"""
 from tkinter import Widget
-from typing import Any, List
+from typing import Any, List, Optional
 
 from pyviews.core import PyViewsError, Setter, Node, XmlNode, InheritedDict, XmlAttr
 from pyviews.expression import parse_expression, is_expression, execute
@@ -55,9 +55,9 @@ class StyleItem:
 class Style(Node):
     """Node for storing config options"""
 
-    def __init__(self, xml_node: XmlNode, node_globals: InheritedDict = None):
+    def __init__(self, xml_node: XmlNode, node_globals: Optional[InheritedDict] = None):
         super().__init__(xml_node, node_globals)
-        self.name = None
+        self.name: Optional[str] = None
         self.items = {}
 
 
@@ -83,8 +83,8 @@ def apply_style_items(node: Style, _: TkRenderingContext):
     attrs = node.xml_node.attrs
     try:
         node.name = next(attr.value for attr in attrs if attr.name == 'name')
-    except StopIteration:
-        raise StyleError('Style name is missing', node.xml_node.view_info)
+    except StopIteration as error:
+        raise StyleError('Style name is missing', node.xml_node.view_info) from error
     node.items = {
         f'{attr.namespace}{attr.name}':
             _get_style_item(node, attr) for attr in attrs if attr.name != 'name'
@@ -117,7 +117,7 @@ def _get_styles(context: TkRenderingContext) -> InheritedDict:
 
 def render_child_styles(node: Style, context: TkRenderingContext):
     """Renders child styles"""
-    render_children(node, context, lambda x, n, ctx: TkRenderingContext({
+    render_children(node, context, lambda x, n, _: TkRenderingContext({
         'parent_node': n,
         'node_globals': InheritedDict(node.node_globals),
         'node_styles': _get_styles(context),
@@ -128,7 +128,7 @@ def render_child_styles(node: Style, context: TkRenderingContext):
 class StylesView(Node):
     """Loads styles from separate file"""
 
-    def __init__(self, master: Widget, xml_node: XmlNode, node_globals: InheritedDict = None):
+    def __init__(self, master: Widget, xml_node: XmlNode, node_globals: Optional[InheritedDict] = None):
         super().__init__(xml_node, node_globals=node_globals)
         self.name = None
         self.master = master